Use semicolon-separated list of fonts for font2
This patch replaces the statically sized array of strings with a single
string which contains a list of fonts to be used as a fallback.
This is useful for specifying multiple fallback fonts with
different sizes in `.Xresources`, for example.

Also the property for `.Xresources` is renamed from `fontalt0` to
`extraFonts`.