Fix segfaults on font reloading
The previous approach was not taking into account that font loading
routine can be called again at some point during runtime. This caused
segfaults when resizing the terminal font and sometimes even randomly by
itself. The problem was that the font library tried to `free()` the
font name string passed to it, but because all font name strings were
actually parts of a single big chunk of memory, attempts to deallocate
any of them (except the first one) resulted in freeing an invalid
pointer.

This commit fixes the issue by creating a dynamically allocated array of
individual strings for each font name instead of just duplicating the
original font list string, replacing semicolons with '\0' and passing
these parts to the font library.